+++ tests/atest/transformers/NormalizeSeparators/test_transformer.py
@@ -51,3 +51,14 @@ class TestNormalizeSeparators(TransformerAcceptanceTest):
 
     def test_skip_documentation(self):
         self.compare(source="test.robot", expected="skip_documentation.robot", config=":skip_documentation=True")
+
+    @pytest.mark.parametrize("indent", [2, 4])
+    @pytest.mark.parametrize("spaces", [2, 4])
+    def test_inline_if(self, spaces, indent):
+        not_modified = indent == 4 and spaces == 4
+        self.compare(
+            source="inline_if.robot",
+            expected=f"inline_if_{indent}indent_{spaces}spaces.robot",
+            config=f" --spacecount {spaces} --indent {indent}",
+            not_modified=not_modified,
+        )